Add an example of the sum (+) type
It might be nicer to define it as a sigma where the first component is
a boolean.  In fact, we should be able to define a single W-type, and
define everything else in terms of that (though it might be nice to be
able to declare things Opaque, so that we don't have to worry about
Brazilians unfolding things when we don't want them to).